+++ internal/buildinfo/buildinfo.go
@@ -2,7 +2,7 @@
 package buildinfo
 
 var (
-	Version = "1.1.0"
+	Version = "1.1.1"
 	Commit  = "unknown"
 	Date    = "unknown"
 )